refactor(agent): promote pending-prompts/global-pause/handoff to runtime services (LifeOps Slice 3) (#8833)
Promote the three cache-backed (verified zero-SQL) LifeOps stores —
pending-prompts, global-pause, handoff — into first-class @elizaos/agent
runtime services, mirroring KnowledgeGraphService.

- Add packages/agent/src/services/{pending-prompts,global-pause,handoff}/
  with the copied store ops + a `class XService extends Service` carrying the
  neutral serviceType literals `eliza_pending_prompts` / `eliza_global_pause`
  / `eliza_handoff` and a `resolveXService(runtime)` accessor.
- Register the three classes in createElizaPlugin() services[] alongside
  KnowledgeGraphService; export them from packages/agent/src/index.ts as
  named re-exports (KG pattern, avoids the broad services barrel / TS2308).
- PA store files re-export the canonical impl from @elizaos/agent (inward-only
  law: agent never imports PA) and add `resolveXStore(runtime)` — prefers the
  registered runtime service (first-wins dedup), falls back to the local
  cache-backed store when absent (both hit the same cache keys).
- PA consumers (pending-prompts/room-policy providers, work-thread evaluator,
  scheduler, scheduled-task action, runtime-wiring) resolve through the helper.
- All runtime string literals (cache keys, serviceType strings) preserved.
