+++ packages/agent/src/services/pending-prompts/service.ts
@@ -1,0 +1,48 @@
+/**
+ * PendingPromptsService — the runtime-owned store of open scheduled-task
+ * prompts, exposed as a registered runtime service.
+ *
+ * Pending prompts are a runtime primitive: any plugin (LifeOps scheduler, the
+ * work-thread evaluator, …) consumes the store via `runtime.getService(...)`
+ * rather than constructing the cache-backed store itself. The service is a thin
+ * factory over the per-runtime {@link PendingPromptsStore}; the store is
+ * cache-backed (no SQL), keyed per room.
+ *
+ * Mirrors `KnowledgeGraphService` (lifecycle + getService accessor).
+ */
+
+import { type IAgentRuntime, Service } from "@elizaos/core";
+import {
+  createPendingPromptsStore,
+  type PendingPromptsStore,
+} from "./store.ts";
+
+export const PENDING_PROMPTS_SERVICE = "eliza_pending_prompts";
+
+export class PendingPromptsService extends Service {
+  static override serviceType = PENDING_PROMPTS_SERVICE;
+
+  override capabilityDescription =
+    "Runtime pending-prompts store: open scheduled-task prompts per room, cache-backed";
+
+  static async start(runtime: IAgentRuntime): Promise<PendingPromptsService> {
+    return new PendingPromptsService(runtime);
+  }
+
+  async stop(): Promise<void> {}
+
+  /** The cache-backed pending-prompts store for this runtime. */
+  getStore(): PendingPromptsStore {
+    return createPendingPromptsStore(this.runtime);
+  }
+}
+
+/**
+ * Resolve the registered {@link PendingPromptsService}. Returns `null` when the
+ * runtime has not registered it (e.g. the "eliza" plugin is absent).
+ */
+export function resolvePendingPromptsService(
+  runtime: IAgentRuntime,
+): PendingPromptsService | null {
+  return runtime.getService<PendingPromptsService>(PENDING_PROMPTS_SERVICE);
+}
