+++ plugins/plugin-personal-assistant/src/lifeops/global-pause/store.ts
@@ -1,122 +1,39 @@
 /**
  * `GlobalPauseStore` — vacation / pause mode singleton.
  *
- * Only ONE pause window can be active at a time. The runner consults
- * `current()` pre-fire; tasks with `respectsGlobalPause: true` skip with
- * reason `global_pause`; tasks with `respectsGlobalPause: false` (emergencies)
- * fire anyway. If `endIso` is set on the active window, the runner reschedules
- * skipped tasks for `endIso` (the user's "resume my routine" moment).
+ * Promoted to a first-class `@elizaos/agent` runtime service in LifeOps Slice 3
+ * (`GlobalPauseService`, serviceType `eliza_global_pause`). The store
+ * implementation and its types now live inward in `@elizaos/agent`; this module
+ * re-exports them and adds {@link resolveGlobalPauseStore}, which prefers the
+ * registered runtime service (first-wins dedup) and falls back to constructing
+ * the cache-backed store directly when the service is absent.
  *
- * Backing storage: runtime cache. Single canonical key.
+ * Backing storage: runtime cache, single canonical key (unchanged).
  */
 
+import {
+  createGlobalPauseStore,
+  type GlobalPauseStore,
+  resolveGlobalPauseService,
+} from "@elizaos/agent";
 import type { IAgentRuntime } from "@elizaos/core";
-import { asCacheRuntime } from "../runtime-cache.js";
 
-export interface GlobalPauseWindow {
-  startIso: string;
-  endIso?: string;
-  reason?: string;
-}
+export {
+  createGlobalPauseStore,
+  GLOBAL_PAUSE_CACHE_KEY,
+  type GlobalPauseStatus,
+  type GlobalPauseStore,
+  type GlobalPauseWindow,
+} from "@elizaos/agent";
 
-export interface GlobalPauseStatus {
-  active: boolean;
-  startIso?: string;
-  endIso?: string;
-  reason?: string;
-}
-
-export interface GlobalPauseStore {
-  set(window: GlobalPauseWindow): Promise<void>;
-  clear(): Promise<void>;
-  current(now?: Date): Promise<GlobalPauseStatus>;
-}
-
-export const GLOBAL_PAUSE_CACHE_KEY = "eliza:lifeops:global-pause:v1";
-
-function isValidIso(value: unknown): value is string {
-  if (typeof value !== "string" || value.length === 0) {
-    return false;
-  }
-  const ms = Date.parse(value);
-  return Number.isFinite(ms);
-}
-
-function normalizeWindow(window: GlobalPauseWindow): GlobalPauseWindow {
-  if (!isValidIso(window.startIso)) {
-    throw new Error(
-      `[global-pause] invalid startIso: ${String(window.startIso)}`,
-    );
-  }
-  if (window.endIso !== undefined && !isValidIso(window.endIso)) {
-    throw new Error(`[global-pause] invalid endIso: ${String(window.endIso)}`);
-  }
-  if (
-    window.endIso !== undefined &&
-    Date.parse(window.endIso) <= Date.parse(window.startIso)
-  ) {
-    throw new Error("[global-pause] endIso must be strictly after startIso");
-  }
-  const normalized: GlobalPauseWindow = { startIso: window.startIso };
-  if (window.endIso !== undefined) {
-    normalized.endIso = window.endIso;
-  }
-  if (typeof window.reason === "string" && window.reason.trim().length > 0) {
-    normalized.reason = window.reason.trim().slice(0, 200);
-  }
-  return normalized;
-}
-
-function isWindowActive(window: GlobalPauseWindow, now: Date): boolean {
-  const startMs = Date.parse(window.startIso);
-  if (!Number.isFinite(startMs) || now.getTime() < startMs) {
-    return false;
-  }
-  if (window.endIso === undefined) {
-    return true;
-  }
-  const endMs = Date.parse(window.endIso);
-  return Number.isFinite(endMs) && now.getTime() < endMs;
-}
-
-export function createGlobalPauseStore(
+/**
+ * Resolve the global-pause store: the registered runtime service when present,
+ * else a directly-constructed cache-backed store. Both read/write the same
+ * canonical cache key, so the fallback is behaviorally identical.
+ */
+export function resolveGlobalPauseStore(
   runtime: IAgentRuntime,
 ): GlobalPauseStore {
-  const cache = asCacheRuntime(runtime);
-
-  return {
-    async set(window: GlobalPauseWindow): Promise<void> {
-      const normalized = normalizeWindow(window);
-      await cache.setCache<GlobalPauseWindow>(
-        GLOBAL_PAUSE_CACHE_KEY,
-        normalized,
-      );
-    },
-    async clear(): Promise<void> {
-      await cache.deleteCache(GLOBAL_PAUSE_CACHE_KEY);
-    },
-    async current(now: Date = new Date()): Promise<GlobalPauseStatus> {
-      const stored = await cache.getCache<GlobalPauseWindow | null>(
-        GLOBAL_PAUSE_CACHE_KEY,
-      );
-      if (!stored || typeof stored !== "object") {
-        return { active: false };
-      }
-      if (!isValidIso(stored.startIso)) {
-        return { active: false };
-      }
-      const active = isWindowActive(stored, now);
-      const status: GlobalPauseStatus = {
-        active,
-        startIso: stored.startIso,
-      };
-      if (stored.endIso !== undefined) {
-        status.endIso = stored.endIso;
-      }
-      if (stored.reason !== undefined) {
-        status.reason = stored.reason;
-      }
-      return status;
-    },
-  };
+  const service = resolveGlobalPauseService(runtime);
+  return service ? service.getStore() : createGlobalPauseStore(runtime);
 }
